refactor: 관리자 리뷰 키워드 통계 로직 안정화 및 키워드 카운트 타입을 Long으로 통일

+++ src/main/java/com/byunsum/ticket_reservation/admin/dto/ReviewStatsResponse.java
@@ -12,7 +12,7 @@ public record ReviewStatsResponse(@Schema(description = "리뷰 총 개수")
                                   Map<String, Long> sentimentCount,
 
                                   @Schema(description = "자주 언급된 키워드")
-                                  Map<String, Integer> topKeywords,
+                                  Map<String, Long> topKeywords,
 
                                   @Schema(description = "평균 감정 점수")
                                   double averageScore,
